Support unmount callout
unmount callout is working as expected now. When removing a pod the
'unmount' call out is call (I thought the spec said unmountdevice).

It is also idempotent, the implementation checks if the given directory
is indeed a mountpoint and only then calls umount. If it is not a mount
point, it simply returns successfully.

Signed-off-by: Roy Golan <rgolan@redhat.com>