Proper thin/thick provisioing behaviour
Motivation
The provisioner should allow thin/thick provisioning of disk, according
to what ovirt supports. Till now the support was through
'ovirtDiskFormat' parameter, and this is now dropped.
If a thin or thick disk is wanted, is is expected using a flag,
ovirtThinProvisioning, true by default.

Modification
- 'ovirtDiskFormat' parameter is dropped from the storage class
- 'ovirtDiskThinProvisioning' parameter, default true is expected.
The api will set the underlying disk format (raw/cow) according to the
storage domain tyep (iscsi/fc/nfs/cluster)

Result
For a thin provision disk, either don't pass nothing or
ovirtDiskThinProvioning=true
If the target ovirt storage domain is ISCSI or FC the disk format type
will be 'cow' and the sparse=true
If the target domain is NFS or Gluster the format is 'raw' and
sparse=true
For ovirtDiskProvisioning=false then the disk format is 'raw' and
sparse=false

fixes: #93
Signed-off-by: Roy Golan <rgolan@redhat.com>