Query the api server to extract node info
Motivation
Few flex actions take a nodeName and can be performed everywhere in the
cluster. Those actions needs to do something on the corespongin VM in
oVirt and this means that the using ovirtVmId from the conf will not
work. For example Attach(volume, nodeName) can be called from the master
and not on the node.

Modification
Flex will query the api to get the node list and extract the systemUUID,
which is the VM id, by the nodeName. It is the equivalent of invoking
this from the cli:
```bash
oc get nodes/$(hostname) -o jsonpath="{.metadata.name} {.status.nodeInfo.systemUUID}"
```

Result
attach/detach/isAttached can be called from every node, to perform
action on every given node.

Fixes: #106
Signed-off-by: Roy Golan <rgolan@redhat.com>